More IRC work, and added Tasks and Misc utils
Further improvements to the IRC classes. They just need message
listeners now so that other parts of the program can add listeners to
certain events, and it needs some way of tracking which users are
online. 

Added a task system, for scheduling future tasks. This will be useful
for polling the Twitch API in regular intervals, polling who are online
in chat and so on.

Last, added some old misc utilities I've used before, such as
KeyPriorityQueues, StackTrace hyperlinks and the JarExplorer.