Added a TwitchAPI and made massive updates to the GUI
TwitchAPI - This package handles communication to Twitch.tv and accesses
their API

GUI - Reworked the GUI completely and utilizes the TwitchAPI to connect
and retreive information from my stream.

Updated documentation and commentating in several places. Also removed a
lot of unnecessary files that had made it to the repository.