Added ObservableUser -> Speedment interaction
When a user is detected in the TwitchIRC channel, a ObservableUser
object is created, wrapping an underlying Speedment User object. Through
that, we can make changes to our ObservableUser (and this see it in
JavaFX) while maintaining updated Speedment objects.

Also did some refactoring in the underlying IrcConnection class, to make
the design more intuitive.