The ability to check if a user is a follower added
I realized that the Twitch API actually has an access point for checking
if a user is a follower or not. This allows me to much easier check for
follower status.

Also, the initialization process may crash if the user enters an invalid
ClientID. Still have not found a work-around to that problem. Might have
to leave it be, since deadline is approaching.