Initialization process almost done
The infinite blocking that occured when the user passed an erroneous
clientID is solved. I had to fork the library I used and add a timeout
to their ServerSocket, but it did the trick. 
The only thing left in the init-process is checking the database for
errors, and integrating it in the program flow (i.e. make it appear on
first launch and then not again.

Added some nice icons to the initialization windows too. Hopefully it is
a bit more fun to look at.

Also, need to start cutting features to make the deadline. Still have to
try it on another computer than this one.